requirements.txt
cos-python-sdk-v5
tencentcloud-sdk-python